requirements.txt
pandas
python-whois
smtplib
python-dotenv
